Adding implementation of lists of articles and lists of articlelists
This also includes a refactoring where we introduce the ListAdapter
and ListFragment abstract classes. These classes are fragments that
contain ListAdapters and the ListAdapters themselves. These classes
introduce automation of setting up lists by introducing template
methods that subclasses need to implement.

This commit also adds a new lifecycle template method to the
BasePresenter class, fragmentInit, which is called after the
presenter has been bound to a fragment and allows for
initialization of the fragment (probably mostly loading of
data) by the presenter.